Moves Files and adds max width to team screen
