+++ src/components/Player/index.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react'
 import { LoadingScreen } from '../ui'
-import { Welcome } from './Welcome'
+import { Welcome } from './ui/Welcome'
 import { CurrentPlayer } from './CurrentPlayer'
 import { Query } from 'react-apollo'
 import { gql } from 'apollo-boost'
